Integrated RevolverOcelot into Current Project
Had to make small changes to Revolver Ocelot (Allow for snake mode when
not using the ocelot button and also a boolean to flip the wheels from
left to right side because I think they are backwards but I'm honestly
not too sure.) Please take a look at all the new code but it should be
ready for just a build and deploy for testing unless I missed something.
Thanks.

-DJ Ramachandran